proyecto_futbol_gq/apps/api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
websockets==12.0
celery==5.3.4
python-socketio==5.10.0
httpx==0.25.1
pandas==2.1.4
numpy==1.26.2
aiofiles==23.2.1
cors==0.1.2